+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/endpoints/MetricsServlet.java
@@ -3,19 +3,17 @@
 package com.yahoo.bard.webservice.web.endpoints;
 
 import static com.yahoo.bard.webservice.config.BardFeatureFlag.UPDATED_METADATA_COLLECTION_NAMES;
-import static javax.ws.rs.core.Response.Status.INTERNAL_SERVER_ERROR;
 
 import com.yahoo.bard.webservice.application.ObjectMappersSuite;
 import com.yahoo.bard.webservice.data.metric.LogicalMetric;
 import com.yahoo.bard.webservice.data.metric.MetricDictionary;
+import com.yahoo.bard.webservice.exception.MetadataExceptionHandler;
 import com.yahoo.bard.webservice.logging.RequestLog;
 import com.yahoo.bard.webservice.logging.blocks.MetricRequest;
 import com.yahoo.bard.webservice.table.LogicalTableDictionary;
-import com.yahoo.bard.webservice.web.ErrorMessageFormat;
-import com.yahoo.bard.webservice.web.apirequest.MetricsApiRequest;
 import com.yahoo.bard.webservice.web.RequestMapper;
-import com.yahoo.bard.webservice.web.RequestValidationException;
 import com.yahoo.bard.webservice.web.ResponseFormatResolver;
+import com.yahoo.bard.webservice.web.apirequest.MetricsApiRequest;
 import com.yahoo.bard.webservice.web.apirequest.MetricsApiRequestImpl;
 
 import com.codahale.metrics.annotation.Timed;
@@ -23,11 +21,11 @@ import com.codahale.metrics.annotation.Timed;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.IOException;
 import java.util.Collection;
 import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
@@ -62,6 +60,7 @@ public class MetricsServlet extends EndpointServlet {
     private final LogicalTableDictionary logicalTableDictionary;
     private final RequestMapper requestMapper;
     private final ResponseFormatResolver formatResolver;
+    private final MetadataExceptionHandler exceptionHandler;
 
     /**
      * Constructor.
@@ -71,6 +70,7 @@ public class MetricsServlet extends EndpointServlet {
      * @param requestMapper  Mapper to change the API request if needed
      * @param objectMappers  JSON tools
      * @param formatResolver  The formatResolver for determining correct response format
+     * @param exceptionHandler  Injection point for handling response exceptions
      */
     @Inject
     public MetricsServlet(
@@ -78,13 +78,15 @@ public class MetricsServlet extends EndpointServlet {
             LogicalTableDictionary logicalTableDictionary,
             @Named(MetricsApiRequest.REQUEST_MAPPER_NAMESPACE) RequestMapper requestMapper,
             ObjectMappersSuite objectMappers,
-            ResponseFormatResolver formatResolver
+            ResponseFormatResolver formatResolver,
+            @Named(MetricsApiRequest.EXCEPTION_HANDLER_NAMESPACE) MetadataExceptionHandler exceptionHandler
     ) {
         super(objectMappers);
         this.metricDictionary = metricDictionary;
         this.logicalTableDictionary = logicalTableDictionary;
         this.requestMapper = requestMapper;
         this.formatResolver = formatResolver;
+        this.exceptionHandler = exceptionHandler;
     }
 
     /**
@@ -115,11 +117,12 @@ public class MetricsServlet extends EndpointServlet {
             @Context final ContainerRequestContext containerRequestContext
     ) {
         Supplier<Response> responseSender;
+        MetricsApiRequest apiRequest = null;
         try {
             RequestLog.startTiming(this);
             RequestLog.record(new MetricRequest("all"));
 
-            MetricsApiRequestImpl apiRequest = new MetricsApiRequestImpl(
+            apiRequest = new MetricsApiRequestImpl(
                     null,
                     formatResolver.apply(format, containerRequestContext),
                     perPage,
@@ -143,23 +146,17 @@ public class MetricsServlet extends EndpointServlet {
                     null
             );
             LOG.debug("Metrics Endpoint Response: {}", response.getEntity());
-            responseSender = () -> response;
-        } catch (RequestValidationException e) {
-            LOG.debug(e.getMessage(), e);
-            responseSender = () -> Response.status(e.getStatus()).entity(e.getErrorHttpMsg()).build();
-        } catch (IOException e) {
-            String msg = String.format("Internal server error. IOException : %s", e.getMessage());
-            LOG.error(msg, e);
-            responseSender = () -> Response.status(INTERNAL_SERVER_ERROR).entity(msg).build();
-        } catch (Error | Exception e) {
-            String msg = ErrorMessageFormat.REQUEST_PROCESSING_EXCEPTION.format(e.getMessage());
-            LOG.info(msg, e);
-            responseSender = () -> Response.status(Response.Status.BAD_REQUEST).entity(msg).build();
+            return response;
+        } catch (Throwable t) {
+            return exceptionHandler.handleThrowable(
+                    t,
+                    Optional.ofNullable(apiRequest),
+                    uriInfo,
+                    containerRequestContext
+            );
         } finally {
             RequestLog.stopTiming(this);
         }
-
-        return responseSender.get();
     }
 
     /**
@@ -182,11 +179,12 @@ public class MetricsServlet extends EndpointServlet {
             @Context final ContainerRequestContext containerRequestContext
     ) {
         Supplier<Response> responseSender;
+        MetricsApiRequest apiRequest = null;
         try {
             RequestLog.startTiming(this);
             RequestLog.record(new MetricRequest(metricName));
 
-            MetricsApiRequest apiRequest = new MetricsApiRequestImpl(
+            apiRequest = new MetricsApiRequestImpl(
                     metricName,
                     null,
                     "",
@@ -208,17 +206,13 @@ public class MetricsServlet extends EndpointServlet {
             String output = objectMappers.getMapper().writeValueAsString(result);
             LOG.debug("Metric Endpoint Response: {}", output);
             responseSender = () -> Response.status(Status.OK).entity(output).build();
-        } catch (RequestValidationException e) {
-            LOG.debug(e.getMessage(), e);
-            responseSender = () -> Response.status(e.getStatus()).entity(e.getErrorHttpMsg()).build();
-        } catch (IOException e) {
-            String msg = String.format("Internal server error. IOException : %s", e.getMessage());
-            LOG.error(msg, e);
-            responseSender = () -> Response.status(INTERNAL_SERVER_ERROR).entity(msg).build();
-        } catch (Error | Exception e) {
-            String msg = ErrorMessageFormat.REQUEST_PROCESSING_EXCEPTION.format(e.getMessage());
-            LOG.info(msg, e);
-            responseSender = () -> Response.status(Response.Status.BAD_REQUEST).entity(msg).build();
+        } catch (Throwable t) {
+            return exceptionHandler.handleThrowable(
+                    t,
+                    Optional.ofNullable(apiRequest),
+                    uriInfo,
+                    containerRequestContext
+            );
         } finally {
             RequestLog.stopTiming(this);
         }