+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/endpoints/TablesServlet.java
@@ -3,39 +3,36 @@
 package com.yahoo.bard.webservice.web.endpoints;
 
 import static com.yahoo.bard.webservice.config.BardFeatureFlag.UPDATED_METADATA_COLLECTION_NAMES;
-import static java.util.AbstractMap.SimpleImmutableEntry;
-import static javax.ws.rs.core.Response.Status.BAD_REQUEST;
-import static javax.ws.rs.core.Response.Status.INTERNAL_SERVER_ERROR;
+
 import static javax.ws.rs.core.Response.Status.OK;
 
 import com.yahoo.bard.webservice.application.ObjectMappersSuite;
 import com.yahoo.bard.webservice.data.config.ResourceDictionaries;
 import com.yahoo.bard.webservice.data.filterbuilders.DruidFilterBuilder;
 import com.yahoo.bard.webservice.data.time.GranularityParser;
+import com.yahoo.bard.webservice.exception.MetadataExceptionHandler;
 import com.yahoo.bard.webservice.logging.RequestLog;
 import com.yahoo.bard.webservice.logging.blocks.TableRequest;
 import com.yahoo.bard.webservice.table.LogicalTable;
 import com.yahoo.bard.webservice.table.LogicalTableDictionary;
 import com.yahoo.bard.webservice.table.resolver.QueryPlanningConstraint;
 import com.yahoo.bard.webservice.util.TableUtils;
-import com.yahoo.bard.webservice.web.ErrorMessageFormat;
 import com.yahoo.bard.webservice.web.RequestMapper;
-import com.yahoo.bard.webservice.web.RequestValidationException;
 import com.yahoo.bard.webservice.web.ResponseFormatResolver;
 import com.yahoo.bard.webservice.web.TableFullViewProcessor;
 import com.yahoo.bard.webservice.web.TableView;
-import com.yahoo.bard.webservice.web.apirequest.TablesApiRequest;
 import com.yahoo.bard.webservice.web.apirequest.HavingGenerator;
+import com.yahoo.bard.webservice.web.apirequest.TablesApiRequest;
 import com.yahoo.bard.webservice.web.apirequest.TablesApiRequestImpl;
 import com.yahoo.bard.webservice.web.util.BardConfigResources;
 
 import com.codahale.metrics.annotation.Timed;
-import com.fasterxml.jackson.core.JsonProcessingException;
 
 import org.joda.time.DateTimeZone;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.util.AbstractMap.SimpleImmutableEntry;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.LinkedHashMap;
@@ -43,8 +40,8 @@ import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
-import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -77,6 +74,7 @@ public class TablesServlet extends EndpointServlet implements BardConfigResource
     private final RequestMapper requestMapper;
     private final GranularityParser granularityParser;
     private final ResponseFormatResolver formatResolver;
+    private final MetadataExceptionHandler exceptionHandler;
 
     /**
      * Constructor.
@@ -86,6 +84,7 @@ public class TablesServlet extends EndpointServlet implements BardConfigResource
      * @param objectMappers  JSON tools
      * @param granularityParser  Helper for parsing granularities
      * @param formatResolver  The formatResolver for determining correct response format
+     * @param exceptionHandler  Injection point for handling response exceptions
      */
     @Inject
     public TablesServlet(
@@ -93,13 +92,15 @@ public class TablesServlet extends EndpointServlet implements BardConfigResource
             @Named(TablesApiRequest.REQUEST_MAPPER_NAMESPACE) RequestMapper requestMapper,
             ObjectMappersSuite objectMappers,
             GranularityParser granularityParser,
-            ResponseFormatResolver formatResolver
+            ResponseFormatResolver formatResolver,
+            @Named(TablesApiRequest.EXCEPTION_HANDLER_NAMESPACE) MetadataExceptionHandler exceptionHandler
     ) {
         super(objectMappers);
         this.resourceDictionaries = resourceDictionaries;
         this.requestMapper = requestMapper;
         this.granularityParser = granularityParser;
         this.formatResolver = formatResolver;
+        this.exceptionHandler = exceptionHandler;
     }
 
     /**
@@ -167,12 +168,12 @@ public class TablesServlet extends EndpointServlet implements BardConfigResource
             @Context UriInfo uriInfo,
             @Context final ContainerRequestContext containerRequestContext
     ) {
-        Supplier<Response> responseSender;
+        TablesApiRequestImpl tablesApiRequestImpl = null;
         try {
             RequestLog.startTiming(this);
             RequestLog.record(new TableRequest(tableName != null ? tableName : "all", "all"));
 
-            TablesApiRequestImpl tablesApiRequestImpl = new TablesApiRequestImpl(
+            tablesApiRequestImpl = new TablesApiRequestImpl(
                     tableName,
                     null,
                     formatResolver.apply(format, containerRequestContext),
@@ -200,19 +201,17 @@ public class TablesServlet extends EndpointServlet implements BardConfigResource
                     null
             );
             LOG.debug("Tables Endpoint Response: {}", response.getEntity());
-            responseSender = () -> response;
-        } catch (RequestValidationException e) {
-            LOG.debug(e.getMessage(), e);
-            responseSender = () -> Response.status(e.getStatus()).entity(e.getErrorHttpMsg()).build();
-        } catch (Error | Exception e) {
-            String msg = ErrorMessageFormat.REQUEST_PROCESSING_EXCEPTION.format(e.getMessage());
-            LOG.info(msg, e);
-            responseSender = () -> Response.status(Response.Status.BAD_REQUEST).entity(msg).build();
+            return response;
+        } catch (Throwable t) {
+            return exceptionHandler.handleThrowable(
+                    t,
+                    Optional.ofNullable(tablesApiRequestImpl),
+                    uriInfo,
+                    containerRequestContext
+            );
         } finally {
             RequestLog.stopTiming(this);
         }
-
-        return responseSender.get();
     }
 
     /**
@@ -242,12 +241,12 @@ public class TablesServlet extends EndpointServlet implements BardConfigResource
             @Context UriInfo uriInfo,
             @Context final ContainerRequestContext containerRequestContext
     ) {
-        Supplier<Response> responseSender;
+        TablesApiRequestImpl tablesApiRequestImpl = null;
         try {
             RequestLog.startTiming(this);
             RequestLog.record(new TableRequest(tableName, grain));
 
-            TablesApiRequestImpl tablesApiRequestImpl = new TablesApiRequestImpl(
+            tablesApiRequestImpl = new TablesApiRequestImpl(
                     tableName,
                     grain,
                     null,
@@ -267,23 +266,17 @@ public class TablesServlet extends EndpointServlet implements BardConfigResource
             Map<String, Object> result = getLogicalTableFullView(tablesApiRequestImpl.getTable(), uriInfo);
             String output = objectMappers.getMapper().writeValueAsString(result);
             LOG.debug("Tables Endpoint Response: {}", output);
-            responseSender = () ->  Response.status(Response.Status.OK).entity(output).build();
-        } catch (RequestValidationException e) {
-            LOG.debug(e.getMessage(), e);
-            responseSender = () ->   Response.status(e.getStatus()).entity(e.getErrorHttpMsg()).build();
-        } catch (JsonProcessingException e) {
-            String msg = ErrorMessageFormat.INTERNAL_SERVER_ERROR_ON_JSON_PROCESSING.format(e.getMessage());
-            LOG.error(msg, e);
-            responseSender = () -> Response.status(INTERNAL_SERVER_ERROR).entity(msg).build();
-        } catch (Error | Exception e) {
-            String msg = ErrorMessageFormat.REQUEST_PROCESSING_EXCEPTION.format(e.getMessage());
-            LOG.info(msg, e);
-            responseSender = () -> Response.status(Response.Status.BAD_REQUEST).entity(msg).build();
+            return Response.status(Response.Status.OK).entity(output).build();
+        } catch (Throwable t) {
+            return exceptionHandler.handleThrowable(
+                    t,
+                    Optional.ofNullable(tablesApiRequestImpl),
+                    uriInfo,
+                    containerRequestContext
+            );
         } finally {
             RequestLog.stopTiming(this);
         }
-
-        return responseSender.get();
     }
 
     /**
@@ -332,12 +325,12 @@ public class TablesServlet extends EndpointServlet implements BardConfigResource
             @Context UriInfo uriInfo,
             @Context final ContainerRequestContext containerRequestContext
     ) {
-        Supplier<Response> responseSender;
+        TablesApiRequestImpl tablesApiRequest = null;
         try {
             RequestLog.startTiming(this);
             RequestLog.record(new TableRequest(tableName, granularity));
 
-            TablesApiRequestImpl tablesApiRequest = new TablesApiRequestImpl(
+            tablesApiRequest = new TablesApiRequestImpl(
                     tableName,
                     granularity,
                     null,
@@ -362,23 +355,17 @@ public class TablesServlet extends EndpointServlet implements BardConfigResource
             Map<String, Object> result = getLogicalTableFullView(tablesApiRequest, uriInfo);
             String output = objectMappers.getMapper().writeValueAsString(result);
             LOG.debug("Tables Endpoint Response: {}", output);
-            responseSender = () ->  Response.status(OK).entity(output).build();
-        } catch (RequestValidationException exception) {
-            LOG.debug(exception.getMessage(), exception);
-            responseSender = () ->   Response.status(exception.getStatus()).entity(exception.getErrorHttpMsg()).build();
-        } catch (JsonProcessingException exception) {
-            String message = ErrorMessageFormat.INTERNAL_SERVER_ERROR_ON_JSON_PROCESSING.format(exception.getMessage());
-            LOG.error(message, exception);
-            responseSender = () -> Response.status(INTERNAL_SERVER_ERROR).entity(message).build();
-        } catch (Error | Exception exception) {
-            String message = ErrorMessageFormat.REQUEST_PROCESSING_EXCEPTION.format(exception.getMessage());
-            LOG.info(message, exception);
-            responseSender = () -> Response.status(BAD_REQUEST).entity(message).build();
+            return Response.status(OK).entity(output).build();
+        } catch (Throwable t) {
+            return exceptionHandler.handleThrowable(
+                    t,
+                    Optional.ofNullable(tablesApiRequest),
+                    uriInfo,
+                    containerRequestContext
+            );
         } finally {
             RequestLog.stopTiming(this);
         }
-
-        return responseSender.get();
     }
 
     /**
@@ -398,12 +385,12 @@ public class TablesServlet extends EndpointServlet implements BardConfigResource
             @Context UriInfo uriInfo,
             @Context final ContainerRequestContext containerRequestContext
     ) {
-        Supplier<Response> responseSender;
+        TablesApiRequestImpl tablesApiRequestImpl = null;
         try {
             RequestLog.startTiming(this);
             RequestLog.record(new TableRequest("all", "all"));
 
-            TablesApiRequestImpl tablesApiRequestImpl = new TablesApiRequestImpl(
+            tablesApiRequestImpl = new TablesApiRequestImpl(
                     null,
                     null,
                     null,
@@ -428,16 +415,17 @@ public class TablesServlet extends EndpointServlet implements BardConfigResource
             Response response = formatResponse(tablesApiRequestImpl, paginatedResult, "tables", null);
 
             LOG.debug("Tables Endpoint Response: {}", response.getEntity());
-            responseSender = () -> response;
-        } catch (Error | Exception e) {
-            String msg = ErrorMessageFormat.REQUEST_PROCESSING_EXCEPTION.format(e.getMessage());
-            LOG.info(msg, e);
-            responseSender = () -> Response.status(Response.Status.BAD_REQUEST).entity(msg).build();
+            return response;
+        } catch (Throwable t) {
+            return exceptionHandler.handleThrowable(
+                    t,
+                    Optional.ofNullable(tablesApiRequestImpl),
+                    uriInfo,
+                    containerRequestContext
+            );
         } finally {
             RequestLog.stopTiming(this);
         }
-
-        return responseSender.get();
     }
 
     /**